move includes to namespaced folder to make deb happy
