+++ examples/ExcitingForceExample.cpp
@@ -20,15 +20,16 @@
 #include <gnuplot-iostream.h>
 #include <iostream>
 #include <limits>
-#include <signal.h>
+#include <csignal>
 #include <string>
 #include <unistd.h>
 #include <vector>
 #include <memory>
 
-#include "config.h"
-#include "FS_Hydrodynamics.hpp"
-#include "LinearIncidentWave.hpp"
+#include <FreeSurfaceHydrodynamics/config.h>
+#include <FreeSurfaceHydrodynamics/FS_Hydrodynamics.hpp>
+#include <FreeSurfaceHydrodynamics/LinearIncidentWave.hpp>
+
 
 void signal_callback_handler(int signum) {
   std::string s = "pkill gnuplot_qt";
