+++ examples/GravityForceExample.cpp
@@ -20,13 +20,14 @@
 #include <gnuplot-iostream.h>
 #include <iostream>
 #include <limits>
-#include <signal.h>
+#include <csignal>
 #include <string>
 #include <unistd.h>
 #include <vector>
 
-#include "config.h"
-#include "FS_Hydrodynamics.hpp"
+#include <FreeSurfaceHydrodynamics/config.h>
+#include <FreeSurfaceHydrodynamics/FS_Hydrodynamics.hpp>
+
 
 void signal_callback_handler(int signum) {
   std::string s = "pkill gnuplot_qt";
